data/crawled_pages/https___firefox_source_docs_mozilla_org_testing_simpletest_html.html

<!doctype html>
<html class="writer-html5" lang="en" data-content_root="../">
 <head>
  <meta charset="utf-8">
  <meta name="viewport" content="width=device-width, initial-scale=1">
  <meta name="viewport" content="width=device-width, initial-scale=1.0">
  <title>SimpleTest framework — Firefox Source Docs  documentation</title>
  <link rel="stylesheet" type="text/css" href="../_static/pygments.css?v=fa44fd50">
  <link rel="stylesheet" type="text/css" href="../_static/css/theme.css?v=19f00094">
  <link rel="stylesheet" type="text/css" href="../_static/graphviz.css?v=fd3f3429">
  <link rel="stylesheet" type="text/css" href="../_static/copybutton.css?v=76b2166b">
  <link rel="stylesheet" type="text/css" href="../_static/custom_theme.css?v=dc5ed815">
  <link rel="stylesheet" type="text/css" href="../_static/design-style.1e8bd061cd6da7fc9cf755528e8ffc24.min.css?v=0a3b3ea7">
  <link rel="shortcut icon" href="../_static/firefox.ico"><!--[if lt IE 9]>
    <script src="../_static/js/html5shiv.min.js"></script>
  <![endif]-->
  <script src="../_static/jquery.js?v=5d32c60e"></script>
  <script src="../_static/_sphinx_javascript_frameworks_compat.js?v=2cd50e6c"></script>
  <script src="../_static/documentation_options.js?v=5929fcd5"></script>
  <script src="../_static/doctools.js?v=9a2dae69"></script>
  <script src="../_static/sphinx_highlight.js?v=dc90522c"></script>
  <script src="../_static/clipboard.min.js?v=a7894cd8"></script>
  <script src="../_static/copybutton.js?v=30646c52"></script>
  <script src="../_static/design-tabs.js?v=36754332"></script>
  <script src="../_static/js/theme.js"></script>
  <link rel="index" title="Index" href="../genindex.html">
  <link rel="search" title="Search" href="../search.html">
 </head>
 <body class="wy-body-for-nav">
  <div class="wy-grid-for-nav">
   <nav data-toggle="wy-nav-shift" class="wy-nav-side">
    <div class="wy-side-scroll">
     <div class="wy-side-nav-search">
      <a href="../index.html" class="icon icon-home"> Firefox Source Docs <img src="../_static/firefox-wordmark.svg" class="logo" alt="Logo"> </a><!-- -- This code is governed by the BSD license ---->
      <div>
       <h3>Quick search</h3>
       <script>
    (function () {
      var cx = "dd12886298f75dbef";
      var gcse = document.createElement("script");
      gcse.async = true;
      gcse.src = "https://cse.google.com/cse.js?cx=" + cx;
      var s = document.getElementsByTagName("script")[0];
      s.parentNode.insertBefore(gcse, s);
    })();
  </script><gcse:search></gcse:search>
      </div>
     </div>
     <div class="wy-menu wy-menu-vertical" data-spy="affix" role="navigation" aria-label="Navigation menu">
      <p class="caption" role="heading"><span class="caption-text">Overview</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="../glossary/index.html">A Glossary of Common Terms</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../overview/index.html">A Quick Guide to Mozilla Applications</a></li>
      </ul>
      <p class="caption" role="heading"><span class="caption-text">Getting Started</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="../setup/index.html">Getting Set Up To Work On The Firefox Codebase</a></li>
      </ul>
      <p class="caption" role="heading"><span class="caption-text">Working On Firefox</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="../contributing/index.html">Working on Firefox</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../bug-mgmt/index.html">Bug Handling</a></li>
      </ul>
      <p class="caption" role="heading"><span class="caption-text">Firefox User Guide</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="../devtools-user/index.html">Firefox DevTools User Docs</a></li>
      </ul>
      <p class="caption" role="heading"><span class="caption-text">Source Code Documentation</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="../mots/index.html">Governance</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../browser/index.html">Firefox Front-end</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../dom/index.html">DOM</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../editor/index.html">Editor</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../layout/index.html">Style system (CSS) &amp; Layout</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../gfx/index.html">Graphics</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../ipc/index.html">Processes, Threads and IPC</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../devtools/index.html">Firefox DevTools Contributor Docs</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../toolkit/index.html">Toolkit</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../js/index.html">SpiderMonkey</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../jsloader/index.html">JS Loader</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../mobile/android/geckoview/index.html">GeckoView</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../mobile/android/fenix/index.html">Fenix</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../mobile/android/focus-android/index.html">Focus for Android</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../dom/bindings/webidl/index.html">WebIDL</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../modules/libpref/index.html">libpref</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../networking/index.html">Networking</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../remote/index.html">Remote Protocols</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../services/index.html">Services</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../permissions/index.html">Permissions</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../uriloader/index.html">File Handling</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../widget/cocoa/index.html">Firefox on macOS</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../widget/windows/index.html">Firefox on Windows</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../toolkit/components/ml/index.html">Firefox AI Runtime</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../accessible/index.html">Accessibility</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../media/index.html">Media Playback</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../code-quality/index.html">Code quality</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../writing-rust-code/index.html">Writing Rust Code</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../rust-components/index.html">Rust Components</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../tools/profiler/index.html">Gecko Profiler</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../performance/index.html">Performance</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../storage/index.html">Database bindings (SQLite, KV, …)</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../xpcom/index.html">XPCOM</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../nspr/index.html">NSPR</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../security/nss/index.html">Network Security Services (NSS)</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../content-security/index.html">Web Security Checks in Gecko</a></li>
      </ul>
      <p class="caption" role="heading"><span class="caption-text">The Firefox Build System</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="../mach/index.html">Mach</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../tools/try/index.html">Pushing to Try</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../build/buildsystem/index.html">Build System</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../taskcluster/index.html">Firefox CI and Taskgraph</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../tools/moztreedocs/index.html">Managing Documentation</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../mozbuild/vendor/index.html">Vendoring Third Party Components</a></li>
      </ul>
      <p class="caption" role="heading"><span class="caption-text">Testing &amp; Test Infrastructure</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="automated-testing/index.html">Automated Testing</a></li>
       <li class="toctree-l1"><a class="reference internal" href="treeherder-try/index.html">Understanding Treeherder Results</a></li>
       <li class="toctree-l1"><a class="reference internal" href="sheriffed-intermittents/index.html">Sheriffed intermittent failures</a></li>
       <li class="toctree-l1"><a class="reference internal" href="tests-for-new-config/index.html">Turning on Firefox tests for a new configuration</a></li>
       <li class="toctree-l1"><a class="reference internal" href="intermittent/index.html">Avoiding intermittent tests</a></li>
       <li class="toctree-l1"><a class="reference internal" href="testing-policy/index.html">Testing Policy</a></li>
       <li class="toctree-l1"><a class="reference internal" href="ci-configs/index.html">Configuration Changes</a></li>
       <li class="toctree-l1"><a class="reference internal" href="browser-chrome/index.html">Browser chrome mochitests</a></li>
       <li class="toctree-l1"><a class="reference internal" href="chrome-tests/index.html">Chrome Tests</a></li>
       <li class="toctree-l1"><a class="reference internal" href="marionette/index.html">Marionette</a></li>
       <li class="toctree-l1"><a class="reference internal" href="geckodriver/index.html">geckodriver</a></li>
       <li class="toctree-l1"><a class="reference internal" href="test-verification/index.html">Test Verification</a></li>
       <li class="toctree-l1"><a class="reference internal" href="webrender/index.html">WebRender Tests</a></li>
       <li class="toctree-l1"><a class="reference internal" href="mochitest-plain/index.html">Mochitest</a></li>
       <li class="toctree-l1"><a class="reference internal" href="xpcshell/index.html">XPCShell tests</a></li>
       <li class="toctree-l1"><a class="reference internal" href="tps/index.html">TPS</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../web-platform/index.html">web-platform-tests</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../gtest/index.html">GTest</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../tools/fuzzing/index.html">Fuzzing</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../tools/sanitizer/index.html">Sanitizer</a></li>
       <li class="toctree-l1"><a class="reference internal" href="perfdocs/index.html">Performance Testing</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../tools/code-coverage/index.html">Code coverage</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../testing-rust-code/index.html">Testing &amp; Debugging Rust Code</a></li>
      </ul>
      <p class="caption" role="heading"><span class="caption-text">Releases &amp; Updates</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="../update-infrastructure/index.html">Mozilla Update Infrastructure</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../update-infrastructure/index.html#watershed-updates">Watershed Updates</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../update-infrastructure/index.html#desupport-updates">Desupport Updates</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../tools/update-verify/index.html">Update Verify</a></li>
      </ul>
      <p class="caption" role="heading"><span class="caption-text">Localization &amp; Internationalization</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="../intl/index.html">Internationalization</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../l10n/index.html">Localization</a></li>
      </ul>
      <p class="caption" role="heading"><span class="caption-text">Firefox and Python</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="../mozbase/index.html">mozbase</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../python/index.html">Using third-party Python packages</a></li>
      </ul>
      <p class="caption" role="heading"><span class="caption-text">Metrics Collected in Firefox</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="../metrics/index.html">Metrics</a></li>
      </ul>
     </div>
    </div>
   </nav>
   <section data-toggle="wy-nav-shift" class="wy-nav-content-wrap">
    <nav class="wy-nav-top" aria-label="Mobile navigation menu">
     <i data-toggle="wy-nav-top" class="fa fa-bars"></i> <a href="../index.html">Firefox Source Docs</a>
    </nav>
    <div class="wy-nav-content">
     <div class="rst-content">
      <!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this file,
   - You can obtain one at http://mozilla.org/MPL/2.0/.  -->
      <div role="navigation" aria-label="Page navigation">
       <ul class="wy-breadcrumbs">
        <li><a href="../index.html" class="icon icon-home" aria-label="Home"></a></li>
        <li class="breadcrumb-item active">SimpleTest framework</li>
        <li class="wy-breadcrumbs-aside"><a href="https://bugzilla.mozilla.org/enter_bug.cgi?product=Developer+Infrastructure&amp;component=Firefox+Source+Docs%3A+Content&amp;short_desc=Documentation+issue+on+testing/simpletest&amp;comment=URL+=+https://firefox-source-docs.mozilla.org/testing/simpletest.html&amp;bug_file_loc=https://firefox-source-docs.mozilla.org/testing/simpletest.html" rel="nofollow">Report an issue</a> / <a href="../_sources/testing/simpletest.rst.txt" rel="nofollow"> View page source</a></li>
       </ul>
       <hr>
      </div>
      <div role="main" class="document" itemscope itemtype="http://schema.org/Article">
       <div itemprop="articleBody">
        <section id="simpletest-framework">
         <h1>SimpleTest framework<a class="headerlink" href="#simpletest-framework" title="Link to this heading">?</a></h1>
         <dl class="js class">
          <dt class="sig sig-object js" id="SimpleTest">
           <em class="property"><span class="k"><span class="pre">class</span></span><span class="w"> </span></em><span class="sig-name descname"><span class="n"><span class="pre">SimpleTest</span></span></span><span class="sig-paren">(</span><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest" title="Link to this definition">?</a>
          </dt>
          <dd>
           <p>SimpleTest framework object.</p>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.doesThrow">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">doesThrow</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">fn</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">name</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.doesThrow" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Check that the function call throws an exception.</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.endMonitorConsole">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">endMonitorConsole</span></span></span><span class="sig-paren">(</span><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.endMonitorConsole" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Stop monitoring console output.</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.executeSoon">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">executeSoon</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">aFunc</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.executeSoon" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Executes a function shortly after the call, but lets the caller continue working (or finish).</p>
             <dl class="field-list simple">
              <dt class="field-odd">
               Arguments<span class="colon">:</span>
              </dt>
              <dd class="field-odd">
               <ul class="simple">
                <li>
                 <p><strong>aFunc</strong> (<em>function</em>) – Function to execute soon.</p></li>
               </ul>
              </dd>
             </dl>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.expectAssertions">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">expectAssertions</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">min</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">max</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.expectAssertions" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Note that the given range of assertions is to be expected. When this function is not called, 0 assertions are expected. When only one argument is given, that number of assertions are expected.</p>
             <p>A test where we expect to have assertions (which should largely be a transitional mechanism to get assertion counts down from our current situation) can call the SimpleTest.expectAssertions() function, with either one or two arguments: one argument gives an exact number expected, and two arguments give a range. For example, a test might do one of the following:</p>
             <p><strong>Examples:</strong></p>
             <div class="highlight-js notranslate">
              <div class="highlight">
               <pre><span></span><span class="c1">// Currently triggers two assertions (bug NNNNNN).</span>
<span class="w">  </span><span class="nx">SimpleTest</span><span class="p">.</span><span class="nx">expectAssertions</span><span class="p">(</span><span class="mf">2</span><span class="p">);</span>

<span class="w">  </span><span class="c1">// Currently triggers one assertion on Mac (bug NNNNNN).</span>
<span class="w">  </span><span class="k">if</span><span class="w"> </span><span class="p">(</span><span class="nx">navigator</span><span class="p">.</span><span class="nx">platform</span><span class="p">.</span><span class="nx">indexOf</span><span class="p">(</span><span class="s2">"Mac"</span><span class="p">)</span><span class="w"> </span><span class="o">==</span><span class="w"> </span><span class="mf">0</span><span class="p">)</span><span class="w"> </span><span class="p">{</span>
<span class="w">    </span><span class="nx">SimpleTest</span><span class="p">.</span><span class="nx">expectAssertions</span><span class="p">(</span><span class="mf">1</span><span class="p">);</span>
<span class="w">  </span><span class="p">}</span>

<span class="w">  </span><span class="c1">// Currently triggers two assertions on all platforms (bug NNNNNN),</span>
<span class="w">  </span><span class="c1">// but intermittently triggers two additional assertions (bug NNNNNN)</span>
<span class="w">  </span><span class="c1">// on Windows.</span>
<span class="w">  </span><span class="k">if</span><span class="w"> </span><span class="p">(</span><span class="nx">navigator</span><span class="p">.</span><span class="nx">platform</span><span class="p">.</span><span class="nx">indexOf</span><span class="p">(</span><span class="s2">"Win"</span><span class="p">)</span><span class="w"> </span><span class="o">==</span><span class="w"> </span><span class="mf">0</span><span class="p">)</span><span class="w"> </span><span class="p">{</span>
<span class="w">    </span><span class="nx">SimpleTest</span><span class="p">.</span><span class="nx">expectAssertions</span><span class="p">(</span><span class="mf">2</span><span class="p">,</span><span class="w"> </span><span class="mf">4</span><span class="p">);</span>
<span class="w">  </span><span class="p">}</span><span class="w"> </span><span class="k">else</span><span class="w"> </span><span class="p">{</span>
<span class="w">    </span><span class="nx">SimpleTest</span><span class="p">.</span><span class="nx">expectAssertions</span><span class="p">(</span><span class="mf">2</span><span class="p">);</span>
<span class="w">  </span><span class="p">}</span>

<span class="w">  </span><span class="c1">// Intermittently triggers up to three assertions (bug NNNNNN).</span>
<span class="w">  </span><span class="nx">SimpleTest</span><span class="p">.</span><span class="nx">expectAssertions</span><span class="p">(</span><span class="mf">0</span><span class="p">,</span><span class="w"> </span><span class="mf">3</span><span class="p">);</span>
</pre>
              </div>
             </div>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.expectChildProcessCrash">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">expectChildProcessCrash</span></span></span><span class="sig-paren">(</span><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.expectChildProcessCrash" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Indicates to the test framework that the current test expects one or more crashes (from plugins or IPC documents), and that the minidumps from those crashes should be removed.</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.expectConsoleMessages">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">expectConsoleMessages</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">testfn</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">msgs</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">continuation</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.expectConsoleMessages" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Run <code class="docutils literal notranslate"><span class="pre">testfn</span></code> synchronously, and monitor its console output.</p>
             <p><code class="docutils literal notranslate"><span class="pre">msgs</span></code> is handled as described above for monitorConsole.</p>
             <p>After <code class="docutils literal notranslate"><span class="pre">testfn</span></code> returns, console monitoring will stop, and <code class="docutils literal notranslate"><span class="pre">continuation</span></code> will be called asynchronously.</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.expectRegisteredServiceWorker">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">expectRegisteredServiceWorker</span></span></span><span class="sig-paren">(</span><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.expectRegisteredServiceWorker" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Indicates to the test framework that this test is expected to leave a service worker registered when it finishes.</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.expectUncaughtException">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">expectUncaughtException</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">aExpecting</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.expectUncaughtException" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Indicates to the test framework that the next uncaught exception during the test is expected, and should not cause a test failure.</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.finish">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">finish</span></span></span><span class="sig-paren">(</span><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.finish" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Finishes the tests. This is automatically called, except when SimpleTest.waitForExplicitFinish() has been invoked.</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.ignoreAllUncaughtExceptions">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">ignoreAllUncaughtExceptions</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">aIgnoring</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.ignoreAllUncaughtExceptions" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Indicates to the test framework that all of the uncaught exceptions during the test are known problems that should be fixed in the future, but which should not cause the test to fail currently.</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.is">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">is</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">a</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">b</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">name</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.is" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Roughly equivalent to ok(Object.is(a, b), name)</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.isExpectingUncaughtException">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">isExpectingUncaughtException</span></span></span><span class="sig-paren">(</span><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.isExpectingUncaughtException" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Returns whether the test has indicated that it expects an uncaught exception to occur.</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.isIgnoringAllUncaughtExceptions">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">isIgnoringAllUncaughtExceptions</span></span></span><span class="sig-paren">(</span><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.isIgnoringAllUncaughtExceptions" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Returns whether the test has indicated that all uncaught exceptions should be ignored.</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.monitorConsole">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">monitorConsole</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">continuation</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">msgs</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">forbidUnexpectedMsgs</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.monitorConsole" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Monitor console output from now until endMonitorConsole is called.</p>
             <p>Expect to receive all console messages described by the elements of <code class="docutils literal notranslate"><span class="pre">msgs</span></code>, an array, in the order listed in <code class="docutils literal notranslate"><span class="pre">msgs</span></code>; each element is an object which may have any number of the following properties:</p>
             <blockquote>
              <div>
               <p>message, errorMessage, sourceName, sourceLine, category: string or regexp lineNumber, columnNumber: number isScriptError, isWarning: boolean</p>
              </div>
             </blockquote>
             <p>Strings, numbers, and booleans must compare equal to the named property of the Nth console message. Regexps must match. Any fields present in the message but not in the pattern object are ignored.</p>
             <p>In addition to the above properties, elements in <code class="docutils literal notranslate"><span class="pre">msgs</span></code> may have a <code class="docutils literal notranslate"><span class="pre">forbid</span></code> boolean property. When <code class="docutils literal notranslate"><span class="pre">forbid</span></code> is true, a failure is logged each time a matching message is received.</p>
             <p>If <code class="docutils literal notranslate"><span class="pre">forbidUnexpectedMsgs</span></code> is true, then the messages received in the console must exactly match the non-forbidden messages in <code class="docutils literal notranslate"><span class="pre">msgs</span></code>; for each received message not described by the next element in <code class="docutils literal notranslate"><span class="pre">msgs</span></code>, a failure is logged. If false, then other non-forbidden messages are ignored, but all expected messages must still be received.</p>
             <p>After endMonitorConsole is called, <code class="docutils literal notranslate"><span class="pre">continuation</span></code> will be called asynchronously. (Normally, you will want to pass <code class="docutils literal notranslate"><span class="pre">SimpleTest.finish</span></code> here.)</p>
             <p>It is incorrect to use this function in a test which has not called SimpleTest.waitForExplicitFinish.</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.ok">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">ok</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">condition</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">name</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.ok" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Something like assert.</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.promiseClipboardChange">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">promiseClipboardChange</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">aExpectedStringOrValidatorFn</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">aSetupFn</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">aFlavor</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">aTimeout</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">aExpectFailure</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">aDontInitializeClipboardIfExpectFailure</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.promiseClipboardChange" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Promise-oriented version of waitForClipboard.</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.promiseFocus">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">promiseFocus</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">aObject</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">expectBlankPage=false</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">aBlurSubframe=false</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.promiseFocus" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>If the page is not yet loaded, waits for the load event. If the page is not yet focused, focuses and waits for the window to be focused. If the current page is ‘<a class="reference external" href="about:blank">about:blank</a>’, then the page is assumed to not yet be loaded. Pass true for expectBlankPage to not make this assumption if you expect a blank page to be present.</p>
             <p>The target object should be specified if it is different than ‘window’. The actual focused window may be a descendant window of aObject.</p>
             <dl class="field-list simple">
              <dt class="field-odd">
               Arguments<span class="colon">:</span>
              </dt>
              <dd class="field-odd">
               <ul class="simple">
                <li>
                 <p><strong>aObject</strong> (<em>Window|browser|BrowsingContext</em>) – Optional object to be focused, and may be any of: window - a window object to focus browser - a &lt;browser&gt;/&lt;iframe&gt; element. The top-level window within the frame will be focused. browsing context - a browsing context containing a window to focus If not specified, defaults to the global ‘window’.</p></li>
                <li>
                 <p><strong>expectBlankPage</strong> (<em>boolean</em>) – True if targetWindow.location is ‘<a class="reference external" href="about:blank">about:blank</a>’.</p></li>
                <li>
                 <p><strong>aBlurSubframe</strong> (<em>boolean</em>) – If true, and a subframe within the window to focus is focused, blur it so that the specified window or browsing context will receive focus events.</p></li>
               </ul>
              </dd>
              <dt class="field-even">
               Returns<span class="colon">:</span>
              </dt>
              <dd class="field-even">
               <p>The browsing context that was focused.</p>
              </dd>
             </dl>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.registerCleanupFunction">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">registerCleanupFunction</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">aFunc</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.registerCleanupFunction" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Register a cleanup/teardown function (which may be async) to run after all tasks have finished, before running the next test. If async (or the function returns a promise), the framework will wait for the promise/async function to resolve.</p>
             <dl class="field-list simple">
              <dt class="field-odd">
               Arguments<span class="colon">:</span>
              </dt>
              <dd class="field-odd">
               <ul class="simple">
                <li>
                 <p><strong>aFunc</strong> (<em>function</em>) – The cleanup/teardown function to run.</p></li>
               </ul>
              </dd>
             </dl>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.registerCurrentTaskCleanupFunction">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">registerCurrentTaskCleanupFunction</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">aFunc</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.registerCurrentTaskCleanupFunction" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Register a cleanup/teardown function (which may be async) to run after the current task has finished, before running the next task. If async (or the function returns a promise), the framework will wait for the promise/async function to resolve.</p>
             <dl class="field-list simple">
              <dt class="field-odd">
               Arguments<span class="colon">:</span>
              </dt>
              <dd class="field-odd">
               <ul class="simple">
                <li>
                 <p><strong>aFunc</strong> (<em>function</em>) – The cleanup/teardown function to run.</p></li>
               </ul>
              </dd>
             </dl>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.registerTaskCleanupFunction">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">registerTaskCleanupFunction</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">aFunc</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.registerTaskCleanupFunction" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Register a cleanup/teardown function (which may be async) to run after each task has finished, before running the next task. If async (or the function returns a promise), the framework will wait for the promise/async function to resolve.</p>
             <dl class="field-list simple">
              <dt class="field-odd">
               Arguments<span class="colon">:</span>
              </dt>
              <dd class="field-odd">
               <ul class="simple">
                <li>
                 <p><strong>aFunc</strong> (<em>function</em>) – The cleanup/teardown function to run.</p></li>
               </ul>
              </dd>
             </dl>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.report">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">report</span></span></span><span class="sig-paren">(</span><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.report" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Makes a test report, returns it as a DIV element.</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.requestCompleteLog">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">requestCompleteLog</span></span></span><span class="sig-paren">(</span><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.requestCompleteLog" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Force all test messages to be displayed. Only applies for the current test.</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.requestFlakyTimeout">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">requestFlakyTimeout</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">reason</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.requestFlakyTimeout" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Request the framework to allow usage of setTimeout(func, timeout) where <code class="docutils literal notranslate"><span class="pre">timeout</span> <span class="pre">&gt;</span> <span class="pre">0</span></code>. This is required to note that the author of the test is aware of the inherent flakiness in the test caused by that, and asserts that there is no way around using the magic timeout value number for some reason.</p>
             <p>Use of this function is <strong>STRONGLY</strong> discouraged. Think twice before using it. Such magic timeout values could result in intermittent failures in your test, and are almost never necessary!</p>
             <dl class="field-list simple">
              <dt class="field-odd">
               Arguments<span class="colon">:</span>
              </dt>
              <dd class="field-odd">
               <ul class="simple">
                <li>
                 <p><strong>reason</strong> (<em>String</em>) – A string representation of the reason why the test needs timeouts.</p></li>
               </ul>
              </dd>
             </dl>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.requestLongerTimeout">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">requestLongerTimeout</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">factor</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.requestLongerTimeout" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Multiply the timeout the parent runner uses for this test by the given factor.</p>
             <p>For example, in a test that may take a long time to complete, using “SimpleTest.requestLongerTimeout(5)” will give it 5 times as long to finish.</p>
             <dl class="field-list simple">
              <dt class="field-odd">
               Arguments<span class="colon">:</span>
              </dt>
              <dd class="field-odd">
               <ul class="simple">
                <li>
                 <p><strong>factor</strong> (<em>Number</em>) – The multiplication factor to use on the timeout for this test.</p></li>
               </ul>
              </dd>
             </dl>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.reset">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">reset</span></span></span><span class="sig-paren">(</span><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.reset" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Resets any state this SimpleTest object has. This is important for browser chrome mochitests, which reuse the same SimpleTest object across a run.</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.runTestExpectingConsoleMessages">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">runTestExpectingConsoleMessages</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">testfn</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">msgs</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.runTestExpectingConsoleMessages" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Wrapper around <code class="docutils literal notranslate"><span class="pre">expectConsoleMessages</span></code> for the case where the test has only one <code class="docutils literal notranslate"><span class="pre">testfn</span></code> to run.</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.showReport">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">showReport</span></span></span><span class="sig-paren">(</span><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.showReport" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Shows the report in the browser</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.todo_is">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">todo_is</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">a</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">b</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">name</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.todo_is" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Copies of is and isnot with the call to ok replaced by a call to todo.</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.toggle">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">toggle</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">el</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.toggle" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Toggle element visibility</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.toggleByClass">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">toggleByClass</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">cls</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">evt</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.toggleByClass" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Toggle visibility for divs with a specific class.</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.waitForCondition">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">waitForCondition</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">aCond</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">aCallback</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">aErrorMsg</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.waitForCondition" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Wait for a condition for a while (actually up to 3s here).</p>
             <dl class="field-list simple">
              <dt class="field-odd">
               Arguments<span class="colon">:</span>
              </dt>
              <dd class="field-odd">
               <ul class="simple">
                <li>
                 <p><strong>aCond</strong> (<em>function</em>) – A function returns the result of the condition</p></li>
                <li>
                 <p><strong>aCallback</strong> (<em>function</em>) – A function called after the condition is passed or timeout.</p></li>
                <li>
                 <p><strong>aErrorMsg</strong> (<em>String</em>) – The message displayed when the condition failed to pass before timeout.</p></li>
               </ul>
              </dd>
             </dl>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.waitForExplicitFinish">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">waitForExplicitFinish</span></span></span><span class="sig-paren">(</span><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.waitForExplicitFinish" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Tells SimpleTest to don’t finish the test when the document is loaded, useful for asynchronous tests.</p>
             <p>When SimpleTest.waitForExplicitFinish is called, explicit SimpleTest.finish() is required.</p>
            </dd>
           </dl>
           <dl class="js staticfunction">
            <dt class="sig sig-object js" id="SimpleTest.waitForFocus">
             <em class="property"><span class="k"><span class="pre">static</span></span><span class="w"> </span></em><span class="sig-prename descclassname"><span class="n"><span class="pre">SimpleTest</span></span><span class="p"><span class="pre">.</span></span></span><span class="sig-name descname"><span class="n"><span class="pre">waitForFocus</span></span></span><span class="sig-paren">(</span><em class="sig-param"><span class="n"><span class="pre">callback</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">aObject</span></span></em>, <em class="sig-param"><span class="n"><span class="pre">expectBlankPage</span></span></em><span class="sig-paren">)</span><a class="headerlink" href="#SimpleTest.waitForFocus" title="Link to this definition">?</a>
            </dt>
            <dd>
             <p>Version of promiseFocus that uses a callback. For compatibility, the callback is passed one argument, the window that was focused. If the focused window is not in the same process, null is supplied.</p>
            </dd>
           </dl>
          </dd>
         </dl>
        </section>
       </div>
      </div>
      <footer>
       <hr>
       <div role="contentinfo">
        <p></p>
       </div> Built with <a href="https://www.sphinx-doc.org/">Sphinx</a> using a <a href="https://github.com/readthedocs/sphinx_rtd_theme">theme</a> provided by <a href="https://readthedocs.org">Read the Docs</a>.
      </footer>
     </div>
    </div>
   </section>
  </div>
  <script>
      jQuery(function () {
          SphinxRtdTheme.Navigation.enable(true);
      });
  </script>
 </body>
</html>
